Remove root, add delegation hashes to the snapshot metadata
According to @JustinCappos in #31, the root.json is no longer
needed in the snapshot metadata, because the workflow will already
have updated the root metadata before the snapshot metadata is
fetched.

In addition, section 5.6 of the Mercury paper describes a need
for delegations to contain hashes in order to protect against
malicous mirrors that may substitute one version of the delegated
metadata with another. This cannot be detected without hashing
of these files.

Closes #31