Remove links to outdated pep 458
The first removed link is supposed to provide additional reasons
for the need of consistent snapshots. Besides pointing to a removed
anchor in an outdated and unmaintained version of pep 485, the spec
already lists the same reasons that are mentioned in the pep.

The second link points to a different (existing) section in the
same outdated document, but gives implementation details that IMO
don't need to be in the spec. If wished we can update the second
link instead of removing it.