delegated targets are authorized by keys in the delegating target roles
This extends the spec to clarify that when searching for the key that
signed a delegated role, that key should only be found in the delegating
target, and not any other role in the delegation chain. This incorporates
language suggested by lukpueh in #58, and the definition of the keys
field from [TAP 3](#57).

Closes #58