Fix inconsistency in root key migration paragraph
Updates an outdated paragraph in section 6.1 that explained how
root keys are rotated by backwards-verifying signatures from the
latest root file available on the repository back to the latest
root file available on the client.

The correct (new) workflow, which is also described in step 1 of
the client application workflow of section 5, performs verification
in a forward-directed manner, i.e. starting at the latest root
on the client and going forward to the latest root on the repo.