Remove 2nd targets rollback attack check
In the client application workflow, remove rollback attack check
for top-level targets file, which is (1) redundant and (2)
prevents recovery from a fast-forward attack.

(1) rollback attacks, via serving older versions of targets or
top-level targets than the previously trusted versions, are already
prevented by step 3.3.3 of the client workflow, where version
numbers of targets and delegated targets in the new snapshot
metadata are asserted to be greater than those in the prior trusted
snapshot metadata.
This, in combination with the 4.1 check that asserts that hashes and
version of the actual targets metadata match the ones in the new
trusted snapshot, makes another version number check, i.e the one
removed in this commit, obsolete.

(2) fast-forward attack recovery, as described in 1.9, works by
having the client remove the trusted timestamp and snapshot
metadata after a non-root key rotation, so that the client can
overcome the version comparison check, and update from a
compromised high version to a recovered lower version.
However, 1.9 does not mention removing trusted targets metadata
after a key rotation. As a consequence, the additional version
number check, removed in this commit, would prevent updating
recovered targets metadata after a fast-forward attack.