Update TARGETPATH and PATHPATTERN recommendations
In addition to recommending the omission of leading directory
separators, this commit adds a recommendation for forward slashes
as preferred directory separators in TARGETPATH and PATHPATTERN.

The rationale is that a TARGETPATH needs to be concatenated with
a mirror's base URL, and as such it might yield unexpected
behavior with non standardized paths. Same goes for Unix
shell-style pattern matching of target paths using PATHPATTERN.

As suggested by @erickt this commit adds a reference to the
"path-relative-URL string" definition from the WHATWG URL spec as
the preferred TARGETPATH style.

This commit also switches to using the term "directory separator"
instead of the prior used "path separator" to disambiguate from
characters that separate paths (e.g. colon on Unix).