Revert "Remove 2nd targets rollback attack check"
This reverts commit 2eccb595628471bcee820f61faa64c2b96a3df60.

2eccb595628471bcee820f61faa64c2b96a3df60 removed rollback attack
check for top-level targets file, which is (1) redundant and (2)
prevents recovery from a fast-forward attack (see commit message
for details).

This commit reverts the change, because the redundancy (1) is
actually desired, so that an attacker does not only have to
compromise timestamp, but snapshot and (delegated) targets too, in
order to launch rollback attacks.

Fast-forward attack recovery (2) shall be fixed in a subsequent
comment.