Add spec version- and release mgmt instructions
- Add gitflow-based version- and release management instructions to
the Versioning section of the README.

- Add links to the "latest stable", "current draft", "new changes
since latest stable" and "release history" of the specification to
the header of the README document.

Note that the version number of the latest stable version of the
specification is no longer shown in the README, which eliminates
the need to update the README with every new release.