Remove TAP 3 + 4 adherence claim
The specification claims that it adheres to TAP 3 and TAP 4,
which is not the case. TAP 3 requires changes to the existing
metadata format (prepared in #57) and TAP 4 adds a new metadata
file (map.json), which is not yet mentioned in the specification
(issue described in #88).

As agreed in the TUF community meeting on Dec 10, 2019, we remove
TAP 3 and TAP 4 (adherence claims), for a formal 1.0.0 release and
add TAP 4 in version 1.1.x and TAP 3 in version 2.x.x.

(see meeting notes about "Versioning"
https://groups.google.com/forum/#!msg/theupdateframework/oPTliXibT_4/0tJ_POLGAwAJ)