Remove 2nd snapshot rollback attack check
This shares the same justification for removal as #65. Step 3.3.1 was made
redundant by #106, which modified the workflow to add 2.2.2, where
updating the timestamp will also check if the new timestamp contains a
snapshot version that is less than the trusted snapshot version.

This, in combination with the 3.1 check that asserts hashes and version of
the actual snapshot metadata match the ones in the new trusted timestamp,
make another version check, i.e, the one removed in this commit, obsolete.