Add a fixed notion of time to the client update workflow
It's possible that metadata may expire during the update cycle in unforeseen
ways. For example, if the timestamp metadata is not expired at the time we
verify it but is expired by the time we are verifying snapshot metadata should
that cause a problem?

This patch asserts not, and codifies this assertion by introducing a fixed
notion of time for the duration of the client update workflow, bounded by
an application defined maximum duration for the update. It is further
recommended that the maximum duration is set to the same time period at
which the repository re-signs the timestamp metadata.

Signed-off-by: Joshua Lock <jlock@vmware.com>