Make section 5 just for client workflow
Adding additional worfklows to what was the "Detailed Workflows" section
would insertion of an additional level of section numbering and
re-numbering the existing steps in the detaile client workflow.

In order to keep things simple and avoid unnecessary re-ordering of the
detailed client workflow, make section 5 deal only with the detailed
client worflow.

Signed-off-by: Joshua Lock <jlock@vmware.com>