Fix METAFILES definition to match the text
The text makes it clear that both "length" and "hashes" are optional and
not dependent on each other. The object format seems to imply that
both should not be included at the same time: DELEGATIONS uses the same
construct to express exactly that.

Express more clearly that the attributes are not dependent on each
other.