Remove workflow maximum duration from update cycle
PR #118 introduced BOTH a fixed notion of time to the update cycle AND an
(arguable misnamed) workflow maximum duration T, which allowed for some
slack in the comparison of expiration time.

The combination of fixed time, expiration time slack, and a change of the
prose describing how to check for a freeze attack resulted in the
specification incorrectly describing a freeze attack check as:

fixed_time = now() + T
expired = expiration > fixed_time

Which results in the freeze attack check requiring that an expiration time
be greater than the start time plus slack, rather than less than the start
time plus slack. That is, that an expiration time in the future could be
incorrectly thought of as expired if it is less than T in the future.

Remove the workflow maximum duration to make things accurate and easier to
reason about.

Signed-off-by: Joshua Lock <jlock@vmware.com>