Require role name uniqueness in delegations
delegations must be unique by role name: otherwise we end up in a
situation where a delegation to a delegated role could be valid and
invalid at the same time, depending on the path or path_hash_prefix.

This matches the treatment in root where roles is an actual dictionary,
ensuring uniqueness by role name.