Check timestamp/snapshot contains snapshot/targets version
The [timestamp.json] states that the timestamp `meta` section MUST only
contain a description of the snapshot.json file. However, [updating the
timestamp role] does not describe when to perform this verification.

Similarly, the [snapshot.json] states that the snapshot `meta` section
MUST contain a description of the targets.json file, which is also not
described in [updating the snapshot role].

This patch explicitly states that these checks should be performed, and
that the metadata should be rejected if it is missing these entries.

[timestamp.json]: https://theupdateframework.github.io/specification/v1.0.30/#file-formats-timestamp
[updating the timestamp role]: https://theupdateframework.github.io/specification/v1.0.30/#update-timestamp
[snapshot.json]: https://theupdateframework.github.io/specification/v1.0.30/#file-formats-snapshot
[updating the snapshot role]: https://theupdateframework.github.io/specification/v1.0.30/#update-snapshot