Update keytype for ecdsa keys and clarify references
We claim that the spec is just documenting the signature _schemes_ from
the reference implementation, but that we define three _keytypes_ within
the spec. This change first updates the _keytypes_ to match the reference
implementation (we have defaulted to a generic "ecdsa" keytype since
https://github.com/secure-systems-lab/securesystemslib/pull/267).

Further, we update the specification to clarify that within we are
documenting the keytypes and schemes from the reference implementation,
and that we recommend implementing these keytypes and schemes as specified.

Signed-off-by: Joshua Lock <jlock@vmware.com>